Create Main.py
This code defines a class GameLibrary that contains methods to add, remove, and display games. In the main() function, we present a simple text-based menu for interacting with the game library. The user can choose different options, such as adding a game, removing a game, displaying the library, or exiting the program.